server_flask/requirements.txt
Flask==2.0.1
tensorflow==2.6.0
Pillow==8.4.0
requests==2.26.0
